Fix wrong background when dialog in the cockpit storage modal view is opened
After we ported the cockpit storage view to a modal, we forgot to adjust
the CSS that handles the backdrop.
